Adds VodCard Adds new CompactCard component Updates Schedule card to use CompactCard Updates CardDetails to use icon passed as prop
