+++ src/components/Card/VodCard.js
@@ -1,14 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import ScheduleCard from './ScheduleCard';
-import PlayIcon from '../../elements/PlayIcon';
+import CompactCard from './CompactCard';
+import play from '../../assets/play.svg';
 
-const VodCard = ({ card, ...props }) => (
-  <div {...props}>
-    <ScheduleCard card={card} icon={<PlayIcon.PlayIconSimple height={20} />} />
-  </div>
-);
+const VodCard = ({ card }) => <CompactCard card={card} icon={<img src={play} alt="play" height={20} />} />;
 
 VodCard.propTypes = {
   card: PropTypes.shape({
